Refactor navigation and modal components
- Renamed `TimerSetupModal` to `IntroModal` and updated references accordingly.
- Introduced `ViewToggleNav` for switching between "bubbles" and "highlights" views in `ThemesPage`.
- Adjusted `WritingTimer` styling for consistency.
- Enhanced `LoadingState` component with updated border color.
- Cleaned up unused code and improved overall structure for better maintainability.
